Move internal annotations to the annotations package
