+++ embrace-android-okhttp3/src/main/java/io/embrace/android/embracesdk/okhttp3/EmbraceCustomPathException.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.okhttp3
 
-import io.embrace.android.embracesdk.InternalApi
+import io.embrace.android.embracesdk.annotation.InternalApi
 import java.io.IOException
 
 /**
