+++ embrace-android-okhttp3/src/main/java/io/embrace/android/embracesdk/okhttp3/swazzle/callback/okhttp3/OkHttpClient.java
@@ -2,7 +2,7 @@ package io.embrace.android.embracesdk.okhttp3.swazzle.callback.okhttp3;
 
 import java.util.List;
 
-import io.embrace.android.embracesdk.InternalApi;
+import io.embrace.android.embracesdk.annotation.InternalApi;
 import io.embrace.android.embracesdk.logging.InternalStaticEmbraceLogger;
 import io.embrace.android.embracesdk.okhttp3.EmbraceOkHttp3ApplicationInterceptor;
 import io.embrace.android.embracesdk.okhttp3.EmbraceOkHttp3NetworkInterceptor;
