+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/Embrace.java
@@ -12,6 +12,7 @@ import androidx.annotation.Nullable;
 import java.util.List;
 import java.util.Map;
 
+import io.embrace.android.embracesdk.annotation.InternalApi;
 import io.embrace.android.embracesdk.config.ConfigService;
 import io.embrace.android.embracesdk.internal.EmbraceInternalInterface;
 import io.embrace.android.embracesdk.logging.InternalEmbraceLogger;
