+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/LogExceptionType.kt
@@ -1,5 +1,7 @@
 package io.embrace.android.embracesdk
 
+import io.embrace.android.embracesdk.annotation.InternalApi
+
 /**
  * Enum representing the type of exception that occurred.
  * NONE is for a native android log, whether have or not an exception.
