+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/annotation/BetaApi.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk
+package io.embrace.android.embracesdk.annotation
 
 /**
  * Classes and methods marked with this annotation are part of Embrace's Beta API that will only function if the app using it is
