+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/EmbraceInternalInterface.kt
@@ -2,8 +2,8 @@ package io.embrace.android.embracesdk.internal
 
 import android.util.Pair
 import io.embrace.android.embracesdk.Embrace
-import io.embrace.android.embracesdk.InternalApi
 import io.embrace.android.embracesdk.LogType
+import io.embrace.android.embracesdk.annotation.InternalApi
 import io.embrace.android.embracesdk.internal.network.http.NetworkCaptureData
 import io.embrace.android.embracesdk.network.EmbraceNetworkRequest
