+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/network/http/EmbraceHttpPathOverride.java
@@ -7,7 +7,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.Locale;
 import java.util.regex.Pattern;
 
-import io.embrace.android.embracesdk.InternalApi;
+import io.embrace.android.embracesdk.annotation.InternalApi;
 import io.embrace.android.embracesdk.logging.InternalStaticEmbraceLogger;
 
 @InternalApi
