+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/payload/Session.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.payload
 
 import com.google.gson.annotations.SerializedName
-import io.embrace.android.embracesdk.InternalApi
+import io.embrace.android.embracesdk.annotation.InternalApi
 import io.embrace.android.embracesdk.session.EmbraceSessionService
 import io.embrace.android.embracesdk.session.MESSAGE_TYPE_START
