+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/utils/ExecutorServiceExtensions.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.utils
 
-import io.embrace.android.embracesdk.InternalApi
+import io.embrace.android.embracesdk.annotation.InternalApi
 import java.util.concurrent.Callable
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Future
